refactor(listener): dissociate EventProducer and TriggerProvider listeners
Fixes gravitee-io/gravitee-alert-engine#177
